requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

Flask
eventlet
flask-cors
flask_socketio
SQLAlchemy
pandas
python_dotenv
pyopenssl
